update to prevent exported mesh naming conflicts:
Blender treats object/datablock names that only differ in case as
NOT equal.  Therefore 'Cube' is not the same as 'cube'.  This doesn't work
for exporters running on windows without internally renaming the mesh
to Cube.001.  Our choice is to display an error message - the
generated scene/meshes will likely NOT be correct.