A bug in Blender maybe.?. debugging the Blender code shows that there isn't a way to extract (via python) shape key data for Ipo curves generated through the action editor. (Key->ipo is always null).
Key channels are supposed to be available via "ipo.curveConsts" which
should return a list of key names.  It doesn't when the Ipo is auto
created through the action editor.

Will report this as a bug and see what comes back.

Also, there doesn't appear to be a link between actions and objects/meshes.
Need to research "action strips" but that would make it a 1+ step process:
    * create action
    * add action to NLA editor
    * convert action to action strip.