Added phyics/collision toggle button to irrb GUI. "Physics.Enabled" attribute value is now based on this button value. When off, no physics/collision attributes are written for mesh nodes.
Added "Physics.Visible" which iwalktest now uses to control whether or not 
the collision mesh is visible."Physics.Ghost" is still exported but not 
used by iwalktest. Decided against using it as a trigger indicator because 
it may be desirable to export an object as "collision w/o restitution" but 
not be a trigger...

irrb-0.4.blend demo is final. "Irrlicth" -> "Irrlicht". really...